setExternalUserId Implementation
Motivation: we want to provide folks the ability to pass the auth hash to `setExternalUserId` while maintaining backwards compatibility. Since we can't overload functions, we implement dynamic variables here by checking what the type of the second argument is. If it is a function, we know it is the callback and no auth hash was provided. If not, it must be a string meaning the callback was omitted and we default to an empty function OR we pass it in as the third argument.