Pass Potential external Id method errors to JS via bridge
Motivation: since we have added user verification to external id setting and modification, we want to surface the error via the function callback by passing it to the JS side in case the error has relevant details.

E.g: "External User Id authentication (auth token) is set to REQUIRED for this application. Please provide an auth token from your backend server or change the setting in the OneSignal dashboard."