Remove native iOS files
Motivation: these files were previously there in order to support non-cocoapods integrations. We are removing it here in the major release and will no longer be maintained moving forward.

The manual linking process is likely done by a small portion of developers and introduces more room for error. Cocoapods is pretty standard at this point and can always be added back later or via SwiftPM in the future.