Small logic bug fix
Motivation: was previously crashing due to passing a string (even though it is a string of a number) to the `sendOutcomeWithValue` function. This change first converts it to a Number type and then checks that it is not NaN which results from trying to cast any non-digit string to a JS Number.