iOS null callback fix for `setExternalUserId`
Motivation: we want to make sure that the `setExternalUserId` works for iOS even if no callback is passed. Thus, we are re-adding the default callback but for iOS only. 

Note: for Android we want it to be nullable temporarily as a workaround for the current issue of the same callback firing multiple times (crashes) due to the `onSuccess` triggering once per channel.

Note: in the future, we intend to make argument non-nullability or nullability consistent across both platforms.