[patch] make handler for promptForPushNotificationsWithUserResponse required in index.d.ts
The original types suggest that the handler is optional. However, running `OneSignal.promptForPushNotificationsWithUserRespons()`(without providing a handler) results in an "Invariant Violation: A valid callback must be supplied" error.

Seeing as all the other methods have their handlers as required, I figured the bug was with the types and not the actual implementation.