Rename helper function
Motivation: this function checks whether the `RNOneSignal` object is non null OR undefined (`!=`) so we should clarify this doesn't check for OneSignal initialization. The new function name should be more clear.

Note: this check exists since PR #485 which provided Expo support (not really support but rather avoid crashing the app in Expo environments)