Fix send outcome functions when no handler is set
The 3 functions related to sending outcomes used to have a default `handler=function(){}` in the function definition if a user does not pass a handler in.

- sendOutcome
- sendUniqueOutcome
- sendOutcomeWithValue

This was removed from the function definition when the SDK was converted to TypeScript in #1312, but the body of the function didn't check and set the handler to an empty function before passing it over the bridge, leading to crashes.